+++ dbt/include/dremio/macros/get_custom_name/get_custom_schema.sql
@@ -15,8 +15,6 @@ limitations under the License.*/
 {% macro dremio__generate_schema_name(custom_schema_name, node) -%}
   {%- set default_schema = target.schema if not is_datalake_node(node)
     else target.root_path -%}
-  {%- set custom_schema_name = custom_schema_name if not is_datalake_node(node)
-    else node.config.root_path -%}
   {{ generate_schema_name_impl(default_schema, custom_schema_name, node) }}
 {%- endmacro %}
 
@@ -25,9 +23,13 @@ limitations under the License.*/
 
       {{ default_schema }}
 
+  {%- elif default_schema == 'no_schema' -%}
+
+      {{ custom_schema_name}}
+
   {%- else -%}
 
-      {{ custom_schema_name }}
+      {{default_schema}}.{{ custom_schema_name }}
 
   {%- endif -%}
 {%- endmacro %}